+++ cpp/libjoynr/joynr-messaging/LibJoynrMessageRouter.cpp
@@ -57,6 +57,7 @@ LibJoynrMessageRouter::LibJoynrMessageRouter(
         std::shared_ptr<IMessagingStubFactory> messagingStubFactory,
         boost::asio::io_service& ioService,
         std::unique_ptr<IMulticastAddressCalculator> addressCalculator,
+        bool persistRoutingTable,
         std::vector<std::shared_ptr<ITransportStatus>> transportStatuses,
         int maxThreads,
         std::unique_ptr<MessageQueue<std::string>> messageQueue,
@@ -65,6 +66,7 @@ LibJoynrMessageRouter::LibJoynrMessageRouter(
                                 std::move(messagingStubFactory),
                                 ioService,
                                 std::move(addressCalculator),
+                                persistRoutingTable,
                                 maxThreads,
                                 std::move(transportStatuses),
                                 std::move(messageQueue),
@@ -310,6 +312,7 @@ void LibJoynrMessageRouter::removeNextHop(
         WriteLocker lock(routingTableLock);
         routingTable.remove(participantId);
     }
+
     saveRoutingTable();
 
     if (!isParentMessageRouterSet()) {